move focal loss to model utils
